RUMM-1100 Move UserInfo into the new core package
