+++ dd-sdk-android/src/main/kotlin/com/datadog/android/log/internal/user/MutableUserInfoProvider.kt
@@ -6,6 +6,7 @@
 
 package com.datadog.android.log.internal.user
 
+import com.datadog.android.core.model.UserInfo
 import com.datadog.tools.annotation.NoOpImplementation
 
 @NoOpImplementation
