--- dd-sdk-android/src/main/kotlin/com/datadog/android/log/internal/user/UserInfo.kt
@@ -1,14 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2016-Present Datadog, Inc.
- */
-
-package com.datadog.android.log.internal.user
-
-internal data class UserInfo(
-    val id: String? = null,
-    val name: String? = null,
-    val email: String? = null,
-    val extraInfo: Map<String, Any?> = emptyMap()
-)
