+++ dd-sdk-android/src/main/kotlin/com/datadog/android/rum/internal/ndk/NdkUserInfoFileStrategy.kt
@@ -8,7 +8,7 @@ package com.datadog.android.rum.internal.ndk
 
 import com.datadog.android.core.internal.domain.FilePersistenceStrategy
 import com.datadog.android.core.internal.privacy.ConsentProvider
-import com.datadog.android.log.internal.user.UserInfo
+import com.datadog.android.core.model.UserInfo
 import com.datadog.android.log.internal.user.UserInfoSerializer
 import java.io.File
 import java.util.concurrent.ExecutorService
