+++ dd-sdk-android/src/test/kotlin/com/datadog/android/log/assertj/LogAssert.kt
@@ -7,8 +7,8 @@
 package com.datadog.android.log.assertj
 
 import com.datadog.android.core.model.NetworkInfo
+import com.datadog.android.core.model.UserInfo
 import com.datadog.android.log.internal.domain.Log
-import com.datadog.android.log.internal.user.UserInfo
 import org.assertj.core.api.AbstractObjectAssert
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.data.Offset
