+++ dd-sdk-android/src/test/kotlin/com/datadog/android/rum/assertj/LongTaskEventAssert.kt
@@ -7,7 +7,7 @@
 package com.datadog.android.rum.assertj
 
 import com.datadog.android.core.model.NetworkInfo
-import com.datadog.android.log.internal.user.UserInfo
+import com.datadog.android.core.model.UserInfo
 import com.datadog.android.rum.internal.domain.scope.isConnected
 import com.datadog.android.rum.model.ErrorEvent
 import com.datadog.android.rum.model.LongTaskEvent
