+++ dd-sdk-android/src/test/kotlin/com/datadog/android/rum/internal/domain/scope/RumSessionScopeTest.kt
@@ -13,8 +13,8 @@ import com.datadog.android.core.internal.CoreFeature
 import com.datadog.android.core.internal.data.NoOpWriter
 import com.datadog.android.core.internal.data.Writer
 import com.datadog.android.core.internal.net.FirstPartyHostDetector
+import com.datadog.android.core.model.UserInfo
 import com.datadog.android.log.internal.logger.LogHandler
-import com.datadog.android.log.internal.user.UserInfo
 import com.datadog.android.rum.GlobalRum
 import com.datadog.android.rum.internal.domain.RumContext
 import com.datadog.android.rum.internal.domain.event.RumEvent
